return from old migrations

+++ db/migrate/20180309083939_remove_position_from_user.rb
@@ -1,9 +1,10 @@
 class RemovePositionFromUser < ActiveRecord::Migration[5.1]
   def up
-  	remove_column :users, :position
+    return if !column_exists?(:users, :position)
+    remove_column :users, :position
   end
 
   def down
-  	add_column :users, :position, :integer
+    add_column :users, :position, :integer
   end
 end
