+++ db/migrate/20180312085319_remove_overriding_percentage.rb
@@ -1,9 +1,10 @@
 class RemoveOverridingPercentage < ActiveRecord::Migration[5.1]
   def up
-  	remove_column :positions, :overriding_percentage
+    return if !column_exists?(:positions, :overriding_percentage)
+    remove_column :positions, :overriding_percentage
   end
 
   def down
-  	add_column :positions, :overriding_percentage, :float
+    add_column :positions, :overriding_percentage, :float
   end
 end
